backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
chromadb==0.4.18
pytesseract==0.3.10
pillow==10.1.0
supabase==2.3.0
httpx>=0.24.0,<0.25.0
python-dateutil==2.8.2
